fix file diff downloader

+++ app/workers/file_diff_downloader.rb
@@ -9,8 +9,8 @@ class FileDiffDownloader
 
   def perform(url, local_path, game_id = nil)
     tmp_file_path = "#{File.dirname(local_path)}/diff_#{rand(1000000)}#{File.extname(local_path)}"
-    `wget "#{url}" -O ""#{tmp_file_path}"`
-    if `cmp "#{local_path}" "#{tmp_file_path}"`.empty?
+    `wget "#{url}" -O "#{tmp_file_path}"`
+    if File.exist?(local_path) && `cmp "#{local_path}" "#{tmp_file_path}"`.empty?
       File.unlink tmp_file_path
     else
       FileUtils.mv tmp_file_path, local_path, :force => true
